+++ google/src/main/scala/org/broadinstitute/dsde/workbench/google/model/GoogleModel.scala
@@ -1,9 +1,0 @@
-package org.broadinstitute.dsde.workbench.google.model
-
-import org.broadinstitute.dsde.workbench.model.{ValueObject, ValueObjectFormat}
-
-case class GoogleProject(value: String) extends ValueObject
-
-object GoogleModelJsonSupport {
-  implicit val GoogleProjectFormat = ValueObjectFormat(GoogleProject)
-}