+++ google/src/main/scala/org/broadinstitute/dsde/workbench/google/model/package.scala
@@ -1,9 +1,0 @@
-package org.broadinstitute.dsde.workbench.google
-
-import org.broadinstitute.dsde.workbench.model.WorkbenchEmail
-
-package object model {
-  def isServiceAccount(email: WorkbenchEmail): Boolean = {
-    email.value.endsWith(".gserviceaccount.com")
-  }
-}