+++ model/src/main/scala/org/broadinstitute/dsde/workbench/model/google/package.scala
@@ -1,0 +1,11 @@
+package org.broadinstitute.dsde.workbench.model
+
+package object google {
+  def isServiceAccount(email: WorkbenchEmail): Boolean = {
+    email.value.endsWith(".gserviceaccount.com")
+  }
+
+  def toAccountName(serviceAccountEmail: WorkbenchEmail): ServiceAccountName = {
+    ServiceAccountName(serviceAccountEmail.value.split("@")(0))
+  }
+}